move javascript interface into own file
